+++ tests/fscripts/demo_json.fs
@@ -1,19 +1,19 @@
 //load from string
-str='{
+var str='{
     "name" : "fscript",
     "age" : 100
 }'
-a=json_load(str)
+var a=json_load(str)
 
 assert(object_get(a, "name")=="fscript");
 assert(object_get(a, "age")==100);
 
 //save to file
-b=json_save(a)
+var b=json_save(a)
 assert(file_write("test.json", b))
 
 //json obj to string
-c=json_save_to_string(a)
+var c=json_save_to_string(a)
 assert(c==str);
 
 //load from file as text
@@ -30,7 +30,3 @@ assert(object_get(a, "name")=="fscript");
 assert(object_get(a, "age")==100);
 
 assert(file_remove("test.json"))
-unset(a)
-unset(b)
-unset(c)
-unset(str)